+++ CHANGES.md
@@ -1,3 +1,8 @@
+TBA
+---
+
+* Support for summary and description (Denis Defreyne)
+
 1.8.0 (2011-06-12)
 ------------------