+++ CHANGES.md
@@ -1,6 +1,8 @@
 TBA
 ---
 
+* Fix: Ensure parsed elements are removed from original arguments when using
+  `:multiple_switches`
 * Ensure anything after `--` is parsed as an argument and not option even
   if prefixed with `/--?/`
 * Performance improvements when making many calls to `Slop#option?` for